lesson9/09_lesson/09_lesson/09_lessonconfig.py
"""Конфигурация для подключения к базе данных"""

import os
from dotenv import load_dotenv

# Загрузка переменных окружения (опционально)
load_dotenv()

# Параметры подключения к БД (замените на свои)
DB_CONFIG = {
    "host": os.getenv("DB_HOST", "localhost"),
    "port": os.getenv("DB_PORT", "5432"),
    "database": os.getenv("DB_NAME", "mydatabase"),
    "user": os.getenv("DB_USER", "myuser"),
    "password": os.getenv("DB_PASSWORD", "mypassword")
}

# Строка подключения
DATABASE_URL = f"postgresql://{DB_CONFIG['user']}:{DB_CONFIG['password']}@{DB_CONFIG['host']}:{DB_CONFIG['port']}/{DB_CONFIG['database']}"