lesson9/09_lesson/09_lesson/09_lessonrequirements.txt
pytest==7.4.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0